backend/requirements.txt
tensorflow>=2.20
tensorflow-datasets>=4.9
numpy>=1.25
Pillow>=10.0
fastapi>=0.110
uvicorn[standard]>=0.23
pydantic-settings>=2.0.0
python-multipart>=0.0.6
matplotlib>=3.8
seaborn>=0.12
scikit-learn>=1.3
importlib-resources>=6.1
etils>=1.1